:fire: Move the hook server into its own package.
